fitness_backend/requirements.txt
absl-py==2.1.0
accelerate==1.3.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
anyio==4.8.0
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.4
asyncio==3.4.3
attrs==24.3.0
axios==0.4.0
babel==2.17.0
beautifulsoup4==4.13.3
bitsandbytes==0.42.0
bleach==6.2.0
cachetools==5.5.1
cattrs==24.1.2
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.2.1
coremltools==8.2
cycler==0.12.1
datasets==3.2.0
debugpy==1.8.12
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.8
diskcache==5.6.3
Django==5.1.5
django-cors-headers==4.6.0
django-embed-video==1.4.10
django-storages==1.14.4
djangorestframework==3.15.2
evaluate==0.4.3
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.17.0
flatbuffers==24.12.23
fonttools==4.55.3
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.9.0
gast==0.6.0
google-api-core==2.24.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpcio==1.69.0
gunicorn==23.0.0
h11==0.14.0
h5py==3.12.1
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.28.1
humanfriendly==10.0
idna==3.10
imageio==2.37.0
ipykernel==6.29.5
ipython==8.32.0
ipywidgets==8.1.5
isoduration==20.11.0
jax==0.5.0
jaxlib==0.5.0
jedi==0.19.2
Jinja2==3.1.5
joblib==1.4.2
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
keras==3.8.0
kiwisolver==1.4.8
lazy_loader==0.4
libclang==18.1.1
llama_cpp_python==0.3.7
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapipe==0.10.18
mistune==3.1.1
ml-dtypes==0.4.1
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
namex==0.0.8
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
notebook==7.3.2
notebook_shim==0.2.4
numpy==1.26.4
onnx==1.17.0
onnxruntime==1.20.1
onnxruntime-tools==1.7.0
onnxruntime_extensions==0.13.0
opencv-contrib-python==4.11.0.86
opt_einsum==3.4.0
optimum==1.10.0
optree==0.14.0
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
peft==0.14.0
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.2.1
proto-plus==1.25.0
protobuf==4.25.5
psutil==6.1.1
psycopg2==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
py3nvml==0.2.7
pyaml==25.1.0
pyarrow==19.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-json-logger==3.2.1
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rouge_score==0.1.2
rpds-py==0.22.3
rsa==4.9
safetensors==0.5.2
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.1
seaborn==0.13.2
Send2Trash==1.8.3
sentencepiece==0.2.0
setuptools==75.8.2
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.1
soupsieve==2.6
sqlparse==0.5.3
stack-data==0.6.3
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
termcolor==2.5.0
terminado==0.18.1
tf_keras==2.18.0
threadpoolctl==3.5.0
tifffile==2025.2.18
tinycss2==1.4.0
tokenizers==0.21.0
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.49.0
trl==0.15.1
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2025.1
uri-template==1.3.0
urllib3==2.3.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
wheel==0.45.1
widgetsnbextension==4.0.13
wrapt==1.17.2
xmltodict==0.14.2
xxhash==3.5.0
yarl==1.18.3
